fix: 헥사고날 아키텍처 폐기 및 도메인 & 엔티티 결합
