+++ build.gradle
@@ -23,6 +23,7 @@ ext {
 
 test {
 	outputs.dir snippetsDir
+	useJUnitPlatform()
 }
 
 asciidoctor {
@@ -74,7 +75,3 @@ spotless {
 		endWithNewline()
 	}
 }
-
-//tasks.named('test') {
-//	useJUnitPlatform()
-//}
