+++ src/main/java/com/ohdab/classroom/domain/classroomInfo/Grade.java
@@ -1,0 +1,13 @@
+package com.ohdab.classroom.domain.classroomInfo;
+
+import lombok.Getter;
+
+@Getter
+public enum Grade {
+    MID_1,
+    MID_2,
+    MID_3,
+    HIGH_1,
+    HIGH_2,
+    HIGH_3;
+}
