--- src/main/java/com/ohdab/entity/main/java/com/ohdab/entity/Grade.java
@@ -1,14 +1,0 @@
-package com.ohdab.entity.main.java.com.ohdab.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum Grade {
-    MID_1,
-    MID_2,
-    MID_3,
-    HIGH_1,
-    HIGH_2,
-    HIGH_3,
-    ;
-}
