+++ src/main/java/com/ohdab/member/controller/MemberController.java
@@ -6,8 +6,8 @@ import com.ohdab.member.controller.request.LoginReq;
 import com.ohdab.member.controller.response.JoinRes;
 import com.ohdab.member.controller.response.LoginRes;
 import com.ohdab.member.service.dto.LoginResDto;
-import com.ohdab.member.service.port.in.JoinUsecase;
-import com.ohdab.member.service.port.in.LoginUsecase;
+import com.ohdab.member.service.usecase.JoinUsecase;
+import com.ohdab.member.service.usecase.LoginUsecase;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
