+++ src/main/java/com/ohdab/member/domain/MemberStatus.java
@@ -1,0 +1,6 @@
+package com.ohdab.member.domain;
+
+public enum MemberStatus {
+    ACTIVE,
+    INACTIVE;
+}
