--- src/main/java/com/ohdab/member/domain/authority/Authority.java
@@ -1,15 +1,0 @@
-package com.ohdab.member.domain.authority;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class Authority {
-
-    private String role;
-
-    @Builder
-    public Authority(String role) {
-        this.role = role;
-    }
-}
