+++ src/main/java/com/ohdab/member/domain/memberinfo/MemberInfo.java
@@ -1,0 +1,15 @@
+package com.ohdab.member.domain.memberinfo;
+
+import javax.persistence.Embeddable;
+import lombok.*;
+
+@Embeddable
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class MemberInfo {
+
+    private String name;
+    private String password;
+}
