+++ src/main/java/com/ohdab/member/exception/DuplicatedMemberException.java
@@ -1,0 +1,17 @@
+package com.ohdab.member.exception;
+
+public class DuplicatedMemberException extends RuntimeException {
+    public DuplicatedMemberException() {}
+
+    public DuplicatedMemberException(String message) {
+        super(message);
+    }
+
+    public DuplicatedMemberException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DuplicatedMemberException(Throwable cause) {
+        super(cause);
+    }
+}
