+++ src/main/java/com/ohdab/member/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package com.ohdab.member.repository;
+
+import com.ohdab.member.domain.Member;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+
+    Optional<Member> findByMemberInfoName(String name);
+}
