+++ src/main/java/com/ohdab/member/service/UserDetailServiceImpl.java
@@ -2,6 +2,8 @@ package com.ohdab.member.service;
 
 import com.ohdab.member.domain.Member;
 import com.ohdab.member.service.helper.MemberHelperService;
+import java.util.List;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
@@ -9,9 +11,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -27,8 +26,8 @@ public class UserDetailServiceImpl implements UserDetailsService {
 
     private UserDetails createUserDetails(Member member) {
         return new User(
-                member.getProfile().getName(),
-                member.getProfile().getPassword(),
+                member.getMemberInfo().getName(),
+                member.getMemberInfo().getPassword(),
                 mapToSimpleGrandAuthority(member));
     }
