--- src/main/java/com/ohdab/member/service/exception/DuplicatedMemberException.java
@@ -1,17 +1,0 @@
-package com.ohdab.member.service.exception;
-
-public class DuplicatedMemberException extends RuntimeException {
-    public DuplicatedMemberException() {}
-
-    public DuplicatedMemberException(String message) {
-        super(message);
-    }
-
-    public DuplicatedMemberException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public DuplicatedMemberException(Throwable cause) {
-        super(cause);
-    }
-}
