--- src/main/java/com/ohdab/member/service/exception/NoMemberException.java
@@ -1,17 +1,0 @@
-package com.ohdab.member.service.exception;
-
-public class NoMemberException extends RuntimeException {
-    public NoMemberException() {}
-
-    public NoMemberException(String message) {
-        super(message);
-    }
-
-    public NoMemberException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public NoMemberException(Throwable cause) {
-        super(cause);
-    }
-}
