+++ src/main/java/com/ohdab/member/service/helper/MemberHelperService.java
@@ -1,8 +1,8 @@
 package com.ohdab.member.service.helper;
 
 import com.ohdab.member.domain.Member;
-import com.ohdab.member.service.exception.NoMemberException;
-import com.ohdab.member.service.port.out.FindMemberPort;
+import com.ohdab.member.exception.NoMemberException;
+import com.ohdab.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public final class MemberHelperService {
 
-    private final FindMemberPort findMemberPort;
+    private final MemberRepository memberRepository;
 
     public Member findExistingMember(String name) {
-        return findMemberPort
-                .findByName(name)
+        return memberRepository
+                .findByMemberInfoName(name)
                 .orElseThrow(() -> new NoMemberException("존재하지 않는 회원입니다."));
     }
 }
