--- src/main/java/com/ohdab/member/service/port/in/JoinUsecase.java
@@ -1,8 +1,0 @@
-package com.ohdab.member.service.port.in;
-
-import com.ohdab.member.service.dto.JoinReqDto;
-
-public interface JoinUsecase {
-
-    void join(JoinReqDto joinReqDto);
-}
