--- src/main/java/com/ohdab/member/service/port/in/LoginUsecase.java
@@ -1,9 +1,0 @@
-package com.ohdab.member.service.port.in;
-
-import com.ohdab.member.service.dto.LoginReqDto;
-import com.ohdab.member.service.dto.LoginResDto;
-
-public interface LoginUsecase {
-
-    LoginResDto login(LoginReqDto loginReqDto);
-}
