+++ src/main/java/com/ohdab/member/service/usecase/JoinUsecase.java
@@ -1,0 +1,8 @@
+package com.ohdab.member.service.usecase;
+
+import com.ohdab.member.service.dto.JoinReqDto;
+
+public interface JoinUsecase {
+
+    void join(JoinReqDto joinReqDto);
+}
