+++ src/main/java/com/ohdab/member/service/usecase/LoginUsecase.java
@@ -1,0 +1,9 @@
+package com.ohdab.member.service.usecase;
+
+import com.ohdab.member.service.dto.LoginReqDto;
+import com.ohdab.member.service.dto.LoginResDto;
+
+public interface LoginUsecase {
+
+    LoginResDto login(LoginReqDto loginReqDto);
+}
