--- src/main/java/com/ohdab/note/domain/student/Student.java
@@ -1,11 +1,0 @@
-package com.ohdab.note.domain.student;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class Student {
-
-    private String name;
-}
