--- src/main/java/com/ohdab/note/domain/workbook/Workbook.java
@@ -1,14 +1,0 @@
-package com.ohdab.note.domain.workbook;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class Workbook {
-
-    private String name;
-    private String description;
-    private int startingNumber;
-    private int endingNumber;
-}
