--- src/main/java/com/ohdab/teacher/domain/group/Grade.java
@@ -1,13 +1,0 @@
-package com.ohdab.teacher.domain.group;
-
-import lombok.Getter;
-
-@Getter
-public enum Grade {
-    MID_1,
-    MID_2,
-    MID_3,
-    HIGH_1,
-    HIGH_2,
-    HIGH_3;
-}
